+++ paddlenlp/transformers/megatronbert/modeling.py
@@ -19,6 +19,11 @@ from paddle import einsum, nn
 
 from .. import PretrainedModel, register_base_model
 from ..activations import get_activation
+from .configuration import (
+    MegatronBert_PRETRAINED_INIT_CONFIGURATION,
+    MegatronBert_PRETRAINED_RESOURCE_FILES_MAP,
+    MegatronBertConfig,
+)
 
 __all__ = [
     "MegatronBertModel",
@@ -46,45 +51,10 @@ class MegatronBertPretrainedModel(PretrainedModel):
 
     """
 
-    pretrained_init_configuration = {
-        "megatronbert-cased": {
-            "attention_probs_dropout_prob": 0.1,
-            "hidden_act": "gelu",
-            "hidden_dropout_prob": 0.1,
-            "hidden_size": 1024,
-            "initializer_range": 0.02,
-            "intermediate_size": 4096,
-            "max_position_embeddings": 512,
-            "num_attention_heads": 16,
-            "num_hidden_layers": 24,
-            "type_vocab_size": 2,
-            "vocab_size": 29056,
-            "pad_token_id": 0,
-        },
-        "megatronbert-uncased": {
-            "attention_probs_dropout_prob": 0.1,
-            "hidden_act": "gelu",
-            "hidden_dropout_prob": 0.1,
-            "hidden_size": 1024,
-            "initializer_range": 0.02,
-            "intermediate_size": 4096,
-            "max_position_embeddings": 512,
-            "num_attention_heads": 16,
-            "num_hidden_layers": 24,
-            "type_vocab_size": 2,
-            "vocab_size": 30592,
-            "pad_token_id": 0,
-        },
-    }
-    pretrained_resource_files_map = {
-        "model_state": {
-            "megatronbert-cased": "http://bj.bcebos.com/paddlenlp/models/transformers/"
-            "megatron-bert/megatronbert-cased/model_state.pdparams",
-            "megatronbert-uncased": "http://bj.bcebos.com/paddlenlp/models/transformers/"
-            "megatron-bert/megatronbert-cased/model_state.pdparams",
-        }
-    }
+    pretrained_init_configuration = MegatronBert_PRETRAINED_INIT_CONFIGURATION
+    pretrained_resource_files_map = MegatronBert_PRETRAINED_RESOURCE_FILES_MAP
     base_model_prefix = "megatronbert"
+    config_class = MegatronBertConfig
 
     def init_weights(self, layer):
         """Initialization hook"""
@@ -107,25 +77,16 @@ class MegatronBertPretrainedModel(PretrainedModel):
 class MegatronBertEmbeddings(nn.Layer):
     """Construct the embeddings from word, position and token_type embeddings."""
 
-    def __init__(
-        self,
-        vocab_size=29056,
-        hidden_size=1024,
-        pad_token_id=0,
-        type_vocab_size=2,
-        max_position_embeddings=512,
-        hidden_dropout_prob=0.1,
-        position_embedding_type="absolute",
-    ):
+    def __init__(self, config: MegatronBertConfig):
         super(MegatronBertEmbeddings, self).__init__()
-        self.word_embeddings = nn.Embedding(vocab_size, hidden_size, padding_idx=pad_token_id)
-        self.position_embeddings = nn.Embedding(max_position_embeddings, hidden_size)
-        self.token_type_embeddings = nn.Embedding(type_vocab_size, hidden_size)
+        self.word_embeddings = nn.Embedding(config.vocab_size, config.hidden_size, padding_idx=config.pad_token_id)
+        self.position_embeddings = nn.Embedding(config.max_position_embeddings, config.hidden_size)
+        self.token_type_embeddings = nn.Embedding(config.type_vocab_size, config.hidden_size)
 
-        self.dropout = nn.Dropout(hidden_dropout_prob)
+        self.dropout = nn.Dropout(config.hidden_dropout_prob)
 
-        self.register_buffer("position_ids", paddle.arange(end=max_position_embeddings).expand((1, -1)))
-        self.position_embedding_type = position_embedding_type
+        self.register_buffer("position_ids", paddle.arange(end=config.max_position_embeddings).expand((1, -1)))
+        self.position_embedding_type = config.position_embedding_type
 
     def forward(
         self, input_ids=None, token_type_ids=None, position_ids=None, inputs_embeds=None, past_key_values_length=0
@@ -157,28 +118,21 @@ class MegatronBertEmbeddings(nn.Layer):
 
 
 class MegatronBertSelfAttention(nn.Layer):
-    def __init__(
-        self,
-        hidden_size=1024,
-        num_attention_heads=16,
-        attention_probs_dropout_prob=0.1,
-        max_position_embeddings=512,
-        position_embedding_type=None,
-    ):
+    def __init__(self, config: MegatronBertConfig):
         super(MegatronBertSelfAttention, self).__init__()
-        self.num_attention_heads = num_attention_heads
-        self.attention_head_size = int(hidden_size / num_attention_heads)
+        self.num_attention_heads = config.num_attention_heads
+        self.attention_head_size = int(config.hidden_size / config.num_attention_heads)
         self.all_head_size = self.num_attention_heads * self.attention_head_size
 
-        self.query = nn.Linear(hidden_size, self.all_head_size)
-        self.key = nn.Linear(hidden_size, self.all_head_size)
-        self.value = nn.Linear(hidden_size, self.all_head_size)
+        self.query = nn.Linear(config.hidden_size, self.all_head_size)
+        self.key = nn.Linear(config.hidden_size, self.all_head_size)
+        self.value = nn.Linear(config.hidden_size, self.all_head_size)
 
-        self.dropout = nn.Dropout(attention_probs_dropout_prob)
-        self.position_embedding_type = position_embedding_type
+        self.dropout = nn.Dropout(config.attention_probs_dropout_prob)
+        self.position_embedding_type = config.position_embedding_type
         if self.position_embedding_type == "relative_key" or self.position_embedding_type == "relative_key_query":
-            self.max_position_embeddings = max_position_embeddings
-            self.distance_embedding = nn.Embedding(2 * max_position_embeddings - 1, self.attention_head_size)
+            self.max_position_embeddings = config.max_position_embeddings
+            self.distance_embedding = nn.Embedding(2 * config.max_position_embeddings - 1, self.attention_head_size)
 
     def transpose_for_scores(self, x):
         new_x_shape = x.shape[:-1] + [self.num_attention_heads, self.attention_head_size]
@@ -232,14 +186,10 @@ class MegatronBertSelfAttention(nn.Layer):
 
 
 class MegatronBertSelfOutput(nn.Layer):
-    def __init__(
-        self,
-        hidden_size=1024,
-        hidden_dropout_prob=0.1,
-    ):
+    def __init__(self, config: MegatronBertConfig):
         super(MegatronBertSelfOutput, self).__init__()
-        self.dense = nn.Linear(hidden_size, hidden_size)
-        self.dropout = nn.Dropout(hidden_dropout_prob)
+        self.dense = nn.Linear(config.hidden_size, config.hidden_size)
+        self.dropout = nn.Dropout(config.hidden_dropout_prob)
 
     def forward(self, hidden_states, residual):
         hidden_states = self.dense(hidden_states)
@@ -248,24 +198,18 @@ class MegatronBertSelfOutput(nn.Layer):
 
 
 class MegatronBertAttention(nn.Layer):
-    def __init__(
-        self,
-        hidden_size=1024,
-        num_attention_heads=16,
-        hidden_dropout_prob=0.1,
-        attention_probs_dropout_prob=0.1,
-        max_position_embeddings=512,
-        position_embedding_type=None,
-    ):
+    def __init__(self, config: MegatronBertConfig):
         super(MegatronBertAttention, self).__init__()
-        self.layer_norm = nn.LayerNorm(hidden_size, epsilon=layer_norm_eps)
+        self.layer_norm = nn.LayerNorm(config.hidden_size, epsilon=layer_norm_eps)
         self.self = MegatronBertSelfAttention(
-            num_attention_heads=num_attention_heads,
-            attention_probs_dropout_prob=attention_probs_dropout_prob,
-            max_position_embeddings=max_position_embeddings,
-            position_embedding_type=position_embedding_type,
+            num_attention_heads=config.num_attention_heads,
+            attention_probs_dropout_prob=config.attention_probs_dropout_prob,
+            max_position_embeddings=config.max_position_embeddings,
+            position_embedding_type=config.position_embedding_type,
         )
-        self.output = MegatronBertSelfOutput(hidden_size=hidden_size, hidden_dropout_prob=hidden_dropout_prob)
+        self.output = MegatronBertSelfOutput(
+            hidden_size=config.hidden_size, hidden_dropout_prob=config.hidden_dropout_prob
+        )
         self.pruned_heads = set()
 
     def forward(self, hidden_states, attention_mask=None):
@@ -301,34 +245,24 @@ class MegatronBertOutput(nn.Layer):
 
 
 class MegatronBertLayer(nn.Layer):
-    def __init__(
-        self,
-        hidden_size=1024,
-        hidden_act="gelu",
-        num_attention_heads=16,
-        hidden_dropout_prob=0.1,
-        attention_probs_dropout_prob=0.1,
-        max_position_embeddings=512,
-        intermediate_size=4096,
-        position_embedding_type=None,
-    ):
+    def __init__(self, config: MegatronBertConfig):
         super(MegatronBertLayer, self).__init__()
         self.seq_len_dim = 1
         self.attention = MegatronBertAttention(
-            hidden_size=hidden_size,
-            num_attention_heads=num_attention_heads,
-            hidden_dropout_prob=hidden_dropout_prob,
-            attention_probs_dropout_prob=attention_probs_dropout_prob,
-            max_position_embeddings=max_position_embeddings,
-            position_embedding_type=position_embedding_type,
+            hidden_size=config.hidden_size,
+            num_attention_heads=config.num_attention_heads,
+            hidden_dropout_prob=config.hidden_dropout_prob,
+            attention_probs_dropout_prob=config.attention_probs_dropout_prob,
+            max_position_embeddings=config.max_position_embeddings,
+            position_embedding_type=config.position_embedding_type,
         )
 
-        self.layer_norm = nn.LayerNorm(hidden_size, epsilon=layer_norm_eps)
+        self.layer_norm = nn.LayerNorm(config.hidden_size, epsilon=layer_norm_eps)
         self.intermediate = MegatronBertIntermediate(
-            hidden_size=hidden_size, intermediate_size=intermediate_size, hidden_act=hidden_act
+            hidden_size=config.hidden_size, intermediate_size=config.intermediate_size, hidden_act=config.hidden_act
         )
         self.output = MegatronBertOutput(
-            intermediate_size, hidden_dropout_prob=hidden_dropout_prob, hidden_size=hidden_size
+            config.intermediate_size, hidden_dropout_prob=config.hidden_dropout_prob, hidden_size=config.hidden_size
         )
 
     def forward(self, hidden_states, attention_mask=None):
@@ -350,38 +284,27 @@ class MegatronBertLayer(nn.Layer):
 
 
 class MegatronBertEncoder(nn.Layer):
-    def __init__(
-        self,
-        hidden_size=1024,
-        hidden_act="gelu",
-        num_attention_heads=16,
-        hidden_dropout_prob=0.1,
-        attention_probs_dropout_prob=0.1,
-        max_position_embeddings=512,
-        intermediate_size=4096,
-        position_embedding_type=None,
-        num_hidden_layers=24,
-    ):
+    def __init__(self, config: MegatronBertConfig):
         super(MegatronBertEncoder, self).__init__()
         self.layer = nn.LayerList(
             [
                 MegatronBertLayer(
-                    hidden_size=hidden_size,
-                    hidden_act=hidden_act,
-                    num_attention_heads=num_attention_heads,
-                    hidden_dropout_prob=hidden_dropout_prob,
-                    attention_probs_dropout_prob=attention_probs_dropout_prob,
-                    max_position_embeddings=max_position_embeddings,
-                    intermediate_size=intermediate_size,
-                    position_embedding_type=position_embedding_type,
+                    hidden_size=config.hidden_size,
+                    hidden_act=config.hidden_act,
+                    num_attention_heads=config.num_attention_heads,
+                    hidden_dropout_prob=config.hidden_dropout_prob,
+                    attention_probs_dropout_prob=config.attention_probs_dropout_prob,
+                    max_position_embeddings=config.max_position_embeddings,
+                    intermediate_size=config.intermediate_size,
+                    position_embedding_type=config.position_embedding_type,
                 )
-                for _ in range(num_hidden_layers)
+                for _ in range(config.num_hidden_layers)
             ]
         )
 
         # The final layer norm. We removed the 1st LN, moved LN to each hidden layer and this one
         # is simply the final LN (Transformer's BERT has it attached to each hidden layer).
-        self.layer_norm = nn.LayerNorm(hidden_size, epsilon=layer_norm_eps)
+        self.layer_norm = nn.LayerNorm(config.hidden_size, epsilon=layer_norm_eps)
 
     def forward(self, hidden_states, attention_mask=None):
         for i, layer_module in enumerate(self.layer):
@@ -423,96 +346,40 @@ class MegatronBertModel(MegatronBertPretrainedModel):
     and refer to the Paddle documentation for all matter related to general usage and behavior.
 
     Args:
-        vocab_size (int):
-            Vocabulary size of `inputs_ids` in `MegatronBertModel`. Also is the vocab size of token embedding matrix.
-            Defines the number of different tokens that can be represented by the `inputs_ids` passed when calling `MegatronBert`.
-        hidden_size (int, optional):
-            Dimensionality of the encoder layer and pooler layer. Defaults to `1024`.
-        pad_token_id (int, optional):
-            The index of padding token in the token vocabulary.
-            Defaults to `0`.
-        type_vocab_size (int, optional):
-            The vocabulary size of `token_type_ids`.
-            Defaults to `2`.
-        hidden_act (str, optional):
-            The non-linear activation function in the feed-forward layer.
-            ``"gelu"``, ``"relu"`` and any other paddle supported activation functions
-            are supported. Defaults to `"gelu"`.
-        attention_probs_dropout_prob (float, optional):
-            The dropout probability used in MultiHeadAttention in all encoder layers to drop some attention target.
-            Defaults to `0.1`.
-        num_attention_heads (int, optional):
-            Number of attention heads for each attention layer in the Transformer encoder.
-            Defaults to `16`.
-        num_hidden_layers (int, optional):
-            Number of hidden layers in the Transformer encoder. Defaults to `24`.
-        max_position_embeddings (int, optional):
-            The maximum value of the dimensionality of position encoding, which dictates the maximum supported length of an input
-            sequence. Defaults to `512`.
-        hidden_dropout_prob (float, optional):
-            The dropout probability for all fully connected layers in the embeddings and encoder.
-            Defaults to `0.1`.
-        intermediate_size (int, optional):
-            Dimensionality of the feed-forward (ff) layer in the encoder. Input tensors
-            to ff layers are firstly projected from `hidden_size` to `intermediate_size`,
-            and then projected back to `hidden_size`. Typically `intermediate_size` is larger than `hidden_size`.
-            Defaults to `4096`.
-        position_embedding_type (str, optional):
-            Type of position embedding. Defaults to "absolute"
-        initializer_range (float, optional):
-            The standard deviation of the normal initializer.
-            Defaults to 0.02.
-
-            .. note::
-                A normal_initializer initializes weight matrices as normal distributions.
-                See :meth:`MegatronBertPretrainedModel.init_weights()` for how weights are initialized in `MegatronBertModel`.
-
+        Args:
+        config (:class:`MegatronBertConfig`):
+            An instance of MegatronBertConfig used to construct MBartModel.
 
     """
 
-    def __init__(
-        self,
-        vocab_size=29056,
-        hidden_size=1024,
-        pad_token_id=0,
-        type_vocab_size=2,
-        hidden_act="gelu",
-        attention_probs_dropout_prob=0.1,
-        num_attention_heads=16,
-        num_hidden_layers=24,
-        max_position_embeddings=512,
-        hidden_dropout_prob=0.1,
-        intermediate_size=4096,
-        position_embedding_type="absolute",
-        initializer_range=0.02,
-    ):
+    def __init__(self, config: MegatronBertConfig):
         super(MegatronBertModel, self).__init__()
 
-        self.num_hidden_layers = num_hidden_layers
-        self.pad_token_id = pad_token_id
-        self.initializer_range = initializer_range
+        self.num_hidden_layers = config.num_hidden_layers
+        self.pad_token_id = config.pad_token_id
+        self.initializer_range = config.initializer_range
         self.embeddings = MegatronBertEmbeddings(
-            vocab_size=vocab_size,
-            hidden_size=hidden_size,
-            pad_token_id=pad_token_id,
-            type_vocab_size=type_vocab_size,
-            max_position_embeddings=max_position_embeddings,
-            hidden_dropout_prob=hidden_dropout_prob,
-            position_embedding_type=position_embedding_type,
+            vocab_size=config.vocab_size,
+            hidden_size=config.hidden_size,
+            pad_token_id=config.pad_token_id,
+            type_vocab_size=config.type_vocab_size,
+            max_position_embeddings=config.max_position_embeddings,
+            hidden_dropout_prob=config.hidden_dropout_prob,
+            position_embedding_type=config.position_embedding_type,
         )
         self.encoder = MegatronBertEncoder(
-            hidden_size=hidden_size,
-            hidden_act=hidden_act,
-            num_attention_heads=num_attention_heads,
-            hidden_dropout_prob=hidden_dropout_prob,
-            attention_probs_dropout_prob=attention_probs_dropout_prob,
-            max_position_embeddings=max_position_embeddings,
-            intermediate_size=intermediate_size,
-            position_embedding_type=position_embedding_type,
-            num_hidden_layers=num_hidden_layers,
+            hidden_size=config.hidden_size,
+            hidden_act=config.hidden_act,
+            num_attention_heads=config.num_attention_heads,
+            hidden_dropout_prob=config.hidden_dropout_prob,
+            attention_probs_dropout_prob=config.attention_probs_dropout_prob,
+            max_position_embeddings=config.max_position_embeddings,
+            intermediate_size=config.intermediate_size,
+            position_embedding_type=config.position_embedding_type,
+            num_hidden_layers=config.num_hidden_layers,
         )
 
-        self.pooler = MegatronBertPooler(hidden_size=hidden_size)
+        self.pooler = MegatronBertPooler(hidden_size=config.hidden_size)
 
         # Initialize weights and apply final processing
         self.apply(self.init_weights)
